Fix printing debug message on unmatched key IDs in getKeyID()
We should not initialize ret to mtc->id to check if it is NULL, because
then after the mapping, ret will always have a non-NULL value making it
not possible to check if the key was unmatched or not. This way we can
print a debug message on unmatched key IDs.

Warned-by: coverity