openpgp-gpg: Switch from internal gpg --list-packets to --verify --status-fd
The GnuPG --list-packets command is for debugging and its interface is
not intended to be used by other tools.

Switch to use --verify with --status-fd, which is a supported interface
with a defined output format. This also simplifies the code parsing.

While this is not an ideal interface, and a dedicated one to inspect a
signature would be better, this is the best we can get right now.

Ref: #1001331