openpgp-gpg: Skip field separator character when parsing --with-colons output
We need to skip the actual field separator or we end up stuck in the
same separator and never match the desired field.

This implied that we returned an empty string that ended up matching
any fingerprint or keyID, bypassing the policy IDs. Although those had
not been really effective anyway due to getKeyID not matching the ID
against the found fingerprint/keyID.