openpgp-gpg: Do not fallback to the match ID on fingerprint/keyID not found
We should always return a fingerprint or a keyID, as that's what we are
going to compare against. Returning the match ID is not helpful because
it both signals we matched something (which might not be the case), and
might be a user ID, which will never match.

Partially fixed with commit a468bcb2b23f114995e0bdd7915409312a6b666c,
but incorrectly restored the fallback behavior, which should have been
removed instead.