openpgp-gpg: When matching on getKeyID match on fingerprint
If we are asked to match on a policy ID, we should check whether it
matches a fingerprint (or keyID) when we find one in the keyring. And
only if not, then match on a UID.

This has meant that the selection and verification policy ID criteria
based on a fingerprint or keyID was not effective until now. This does
not seem like a serious issue though, as the policy needed to be found
based on the fingerprint/keyID of the signature, the origin and type
checks were in place and working, and the signature needed to verify
successfully based on the specific policy keyring.