Check whether the origin ID is in the keyring if no match ID is specified
We should check whether the origin ID can be found in the keyring, as
that is the fingerprint/keyID for the signature that selected the
current policy.

It's not essential, as the actual signature verification will be the
final decision point, but it means we might be able to discard this
policy as invalid otherwise, and use another one instead.