openpgp-gpg: Move and initialize fpr declaration outside of the getline loop
The loop uses this variable on subsequent lines when matched to return
the fingerprint found, so we need to widen its scope. In addition we
need to initialize it to NULL, even though the code takes care of not
using ret when it is NULL, and only set ret to fpr, on the specific
state.

Fixes: commit 29f80390bebba1deb304ac58e0acc0644cb130eb
Warned-by: gcc
Changelog: silent